Clean test cases
