+++ tests/ORM/Driver/Postgres/RefersToRelationRenamedFieldsTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Cycle\ORM\Tests\Driver\Postgres;
 
-class RefersToRelationTestRenamedFields extends \Cycle\ORM\Tests\RefersToRelationTestRenamedFields
+class RefersToRelationRenamedFieldsTest extends \Cycle\ORM\Tests\RefersToRelationRenamedFieldsTest
 {
     public const DRIVER = 'postgres';
 }
