+++ tests/ORM/RefersToRelationRenamedFieldsTest.php
@@ -1,0 +1,80 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests;
+
+use Cycle\ORM\Mapper\Mapper;
+use Cycle\ORM\Relation;
+use Cycle\ORM\Schema;
+use Cycle\ORM\Tests\Fixtures\Comment;
+use Cycle\ORM\Tests\Fixtures\User;
+
+abstract class RefersToRelationRenamedFieldsTest extends RefersToRelationTest
+{
+    public function setUp(): void
+    {
+        BaseTest::setUp();
+
+        $this->makeTable('user', [
+            'id'         => 'primary',
+            'email'      => 'string',
+            'balance'    => 'float',
+            'last_comment_id_field' => 'integer,nullable'
+        ]);
+
+        $this->makeTable('comment', [
+            'comment_id' => 'primary',
+            'user_id' => 'integer',
+            'message' => 'string'
+        ], [
+            'user_id' => ['table' => 'user', 'column' => 'id']
+        ]);
+
+        $this->orm = $this->withSchema(new Schema([
+            User::class    => [
+                Schema::ROLE        => 'user',
+                Schema::MAPPER      => Mapper::class,
+                Schema::DATABASE    => 'default',
+                Schema::TABLE       => 'user',
+                Schema::PRIMARY_KEY => 'id',
+                Schema::COLUMNS     => ['id', 'email', 'balance', 'comment_id' => 'last_comment_id_field'],
+                Schema::TYPECAST    => ['id' => 'int'],
+                Schema::RELATIONS   => [
+                    'lastComment' => [
+                        Relation::TYPE   => Relation::REFERS_TO,
+                        Relation::TARGET => Comment::class,
+                        Relation::SCHEMA => [
+                            Relation::CASCADE   => true,
+                            Relation::INNER_KEY => 'comment_id',
+                            Relation::OUTER_KEY => 'id'
+                        ],
+                    ],
+                    'comments'    => [
+                        Relation::TYPE   => Relation::HAS_MANY,
+                        Relation::TARGET => Comment::class,
+                        Relation::SCHEMA => [
+                            Relation::CASCADE   => true,
+                            Relation::INNER_KEY => 'id',
+                            Relation::OUTER_KEY => 'user_id',
+                        ],
+                    ],
+
+                ]
+            ],
+            Comment::class => [
+                Schema::ROLE        => 'comment',
+                Schema::MAPPER      => Mapper::class,
+                Schema::DATABASE    => 'default',
+                Schema::TABLE       => 'comment',
+                Schema::PRIMARY_KEY => 'id',
+                Schema::COLUMNS     => [
+                    'id' => 'comment_id',
+                    'user_id',
+                    'message',
+                ],
+                Schema::SCHEMA      => [],
+                Schema::RELATIONS   => [],
+            ],
+        ]));
+    }}
